Add files via upload
Nashville housing data is the subject of this dataset. This dataset provides details about each home in Nashville, including price, area, bathrooms, and bedrooms. This information describes many sample collections of homes and their attributes. There are 26 variables in 22651 samples in this data (features). The goal of reporting on this dataset is to determine the factors that Nashville home prices are most influenced by. Therefore, a corporation might think about those elements and take them into account when making decisions about purchasing new homes and making real estate investments. This report discusses the entire data analysis process, the findings, and the next steps the organization should take. The entire process of analysis is explained, along with recommendations for new buying tactics and information on whether the house is overpriced or underpriced.